% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_get_files.R
\name{bt_get_files}
\alias{bt_get_files}
\title{bt_get_files}
\usage{
bt_get_files(folder, pattern = "", date = FALSE)
}
\arguments{
\item{folder}{Directory containing files of interest.}

\item{pattern}{Optional pattern to use in file searching.}

\item{date}{Do file names contain a date which should be removed?}
}
\value{
Named list of files to be read.
}
\description{
Function which creates a named list of files in a specified
  directory. Names are trimmed versions of file names, contents of the list
  are the file names themselves. Can be easily piped into
  \code{purrr::map(~read.csv(.))} to create named list of data frames.
}
\references{
None.
}
\seealso{
https://github.com/hancockinformatics/BacterialToolkit
}
