% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_gtf_cleaner.R
\name{bt_gtf_cleaner}
\alias{bt_gtf_cleaner}
\title{bt_gtf_cleaner}
\usage{
bt_gtf_cleaner(gtf_file)
}
\arguments{
\item{gtf_file}{Path to GTF/GFF file to be cleaned}
}
\value{
Tidy data frame of annotations
}
\description{
Given an input GTF file (Pseudomonas aeruginosa), separates and cleans
columns, returning a clean and tidy data frame. Only returns locus tag, gene
name, description, start, end, and strand columns. Only supports PAO1, PA14,
and LESB58. Uses a single regex to match and extract locus tag for all three
strains.
}
\author{
Travis Blimkie
}
