% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_theme.R
\name{bt_theme}
\alias{bt_theme}
\title{bt_theme}
\usage{
bt_theme(baseSize = 16, baseFamily = "", grid = FALSE)
}
\arguments{
\item{baseSize}{Base font size applied to all text in the plot}

\item{baseFamily}{Base font type applies to all text in the plot}

\item{grid}{Determines if grid lines should be drawn. Defaults to FALSE. When
TRUE, only "major" grid lines are included.}
}
\value{
Theme object
}
\description{
Custom theme that can be applied to ggplot2 plots. Increases
  base font size, sets a white background, grey grid, and black border.
}
\examples{
\dontrun{
  ggplot(mtcars, aes(cyl, mpg)) + geom_point() + tr_theme(grid = TRUE)
}

}
\references{
None.
}
\seealso{
\url{https://www.github.com/hancockinformatics/BacterialToolkit}
}
