% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_tidy_gage.R
\name{bt_tidy_gage}
\alias{bt_tidy_gage}
\title{bt_tidy_gage}
\usage{
bt_tidy_gage(gage_result, qval = 0.1)
}
\arguments{
\item{gage_result}{Results object from Gage enrichment.}

\item{qval}{Desired cutoff for significance.}
}
\value{
Tidy data frame containing results of Gage enrichment.
}
\description{
Takes a results object from Gage and munges it into a tidy data
frame Also filters based on q-value, with a default of 0.1 from the Gage
documentation.
}
\references{
None.
}
\seealso{
https://github.com/hancockinformatics/BacterialToolkit
}
\author{
Travis Blimkie
}
